Move getRelatedTerms function to helpers.js
